p2p : Rename basic server to test server
